- added progress bar - retrieving movies logic moved to fragment
