+++ app/src/main/res/layout/fragment_movies_list.xml
@@ -6,6 +6,7 @@
     android:layout_height="match_parent"
     tools:context=".activity.MainActivity">
 
+
     <Button
         android:id="@+id/day_night_mode"
         android:layout_width="42dp"
@@ -13,7 +14,19 @@
         android:background="@android:color/transparent"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        android:drawableStart="@drawable/light_mode" />
+        android:drawableStart="@drawable/light_mode"
+        android:visibility="gone"/>
+
+    <ProgressBar
+        android:id="@+id/progress_bar"
+        style="?android:progressBarStyle"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintStart_toEndOf="@+id/day_night_mode"
+        app:layout_constraintEnd_toStartOf="@id/add_new"
+        />
 
 
     <ImageButton
@@ -22,7 +35,8 @@
         android:layout_height="wrap_content"
         android:src="@drawable/add_new"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
+        app:layout_constraintTop_toTopOf="parent"
+        android:visibility="gone">
     </ImageButton>
